restructure: "app" -> "ui" for almost every presentational component
